app/requirements.txt
asgiref==3.8.1
blinker==1.8.2
Brotli==1.1.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
cssselect2==0.7.0
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.1.3
django-appconf==1.1.0
django-bootstrap-modal-forms==3.0.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-compressor==4.5.1
django-filter==24.2
django-formset==1.4.5
django-js-asset==2.2.0
django-plotly-dash==2.3.1
django-safedelete==1.3.2
django-weasyprint==2.4.0
django-widget-tweaks==1.5.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.1
dpd_components==0.1.0
Flask==3.0.3
fonttools==4.53.0
gunicorn==23.0.0
html5lib==1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
Pillow==10.1.0
plotly==5.22.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
pytz==2024.1
rcssmin==1.1.2
rjsmin==1.2.2
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.2
tenacity==8.3.0
tinycss2==1.3.0
tzdata==2024.1
weasyprint==62.1
webencodings==0.5.1
Werkzeug==3.0.3
wheel==0.45.1
wkhtmltopdf==0.2
zopfli==0.2.3
